Add starting ROBE, adjust player/NPC structs, etc.
Decided to provide a ROBE as the only starting item and to allow up to 4
heavy items to be carried at once. Also updated the data members of the
player_t and npc_t structs, added MAX_DEPTH and MAX_LEVEL constants (to
be utilized later), and modified "adjust_minor_stats" which has been
renamed as "set_player_minor_stats".